refactor(hir-ty): rename inference implementation
