+++ crates/hir-ty/src/lib.rs
@@ -7,7 +7,7 @@
 
 pub mod db;
 pub mod display;
-mod type_infer;
+mod infer;
 mod type_system;
 
 pub use type_system::{Compatibility, Member, Type, TypeDiagnostic, TypeSystem};
